pyproject.toml
[project]
name = "manim-chatbot"
version = "0.1.0"
description = "Chat-driven Manim script generator"
authors = [
    {name = "Abhay Mishra grabhaymishra@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "transformers (>=4.51.3,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "dotenv (>=0.0.4,<0.10.0)",
    "supabase (>=2.15.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
